+++ src/store/feed/reducer.js
@@ -17,11 +17,6 @@ export default function feedSliceReducer(state = initialState, action) {
         posts: [...action.payload],
       };
     }
-    case "ADD_HEART":
-      return {
-        ...state,
-        posts: { likes: action.payload },
-      };
 
     case "POST_CREATED":
       return {
